requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Werkzeug>=2.0 # For password hashing, usually installed with Flask
python-dotenv>=0.19 # For loading .env files
PyMySQL>=1.0 # MySQL driver for SQLAlchemy
